Fix two bugs.  Neither bug affects results.
- #2654 computation of maxloops in oasis_mpi_reducelists.  The original
implementation was incorrect.  This does not affect results, but could
have resulted in integer overflow for high pe counts.  The new implementation
should avoid the problem and reduce the number of searches at the same time.

- #2655 array bounds overflow in cplfind.  With some compilers and array
bounds checking on, a combined check in one statement created an array bounds
overflow at the point in the do loop where the loop was exiting.  The
implementation was refactored to separate the combined check into
two separate checks.


Former-commit-id: fba830ffa4b5c2f47c443bfebf37afabcfe372e0